Add types to serialization
