+++ src/BlazorDatasheet.Serialization.Json/Models/DataRegionModel.cs
@@ -2,12 +2,12 @@
 
 namespace BlazorDatasheet.Serialization.Json.Models;
 
-public class DataRegion<T>
+public class DataRegionModel<T>
 {
     [JsonPropertyName("sqref")] public string RegionString { get; set; }
     [JsonPropertyName("v")] public T Value { get; set; }
 
-    public DataRegion(string regionString, T value)
+    public DataRegionModel(string regionString, T value)
     {
         RegionString = regionString;
         Value = value;
